Initial support for asynchronous tasks using Celery, -- not yet fully integrated into EDD configuration.
At this stage, this only includes directions and configuration files for running RabbitMQ, Celery, and Flower, and for demonizing RabbitMQ and Celery on Debian and OSX. These directions should be sufficient for running the servers in production on a single host, but note many lower-priority TODO's and the known need for additional formatting in README.md.  Tested basic task scheduling from the command line and control via the web interfaces.  Full Django integration not included at this stage to avoid breaking edd-test during experimentation.