Fixing a typo.
Note that there is a potential issue with adding "modified at" timestamps:  Every time a developer switches between branches, some of the file timestamps will be moved to the current date.  So these timestamps are only useful to developers insofar as they inform them that a file is newer than the last time they did a branch change.  Still more useful than a global "compile time" date ... but not static enough that we don't still need to filter it out from checkins.