Consolidating GraphHelperMethods, barGraphHelperMethods, createMultiAxisLine, and lineGraphHelperMethods into EDDGraphingTools, a Typescript-friendly file using its own EDDGraphingTools namespace.  Altering the Study and Import pages to use it.
Consolidating and Typescript-ifying groupedBarMeasurement, StudyGraphing, and StudyGraphingHelperMethods into Study-Data, which is now under the StudyDataPage namespace (instead of just 'StudyData') to avoid confusion.

Rearranging and rewriting most of the logic in Study-Data.  Now the line graph, various bar graphs, and table only redraw when each of them is brought to the foreground AND is changed.  This speeds the page up considerably.

Consolidating the table controls, graph filter controls, and action buttons into one modular bar above the filtering section.  Still need to root out an exception that happens in the table, and an issue with the Measurement-based bar graphs, and some other details.

Also need to re-integrate the new 'show empty assays' button.