Removing HTML template script and fix view switch
There was an inline script in the HTML template for the study-data
page, which would trigger a button click at the end of every AJAX
request, and that button click would trigger another AJAX request,
causing an infinite loop. The same inline script was included in
Study-Data.ts, and is now using the refactored display function and
does not trigger an AJAX request.