Adding autocomplete-aware auto-fill
The EDD autocomplete widgets now trigger a 'autochange' event, receiving
additional arguments for the value of the display input and the value of
the hidden input. The import page listens for this event, and auto-fills
compartment and unit inputs following the changed input; aborting if a
following input was previously set by a user (do not overwrite user
values).