Remove use of Channels parent class functionality
The parent classes of the generic Channels websocket consumers can auto-join
the consumer to a Channel Group on connect. However, existing code was trying
to use the authenticated user to generate a group name, and the consumer scope
does not have the authenticated user available at connect-time. Therefore, the
consumers must duplicate functionality of the parent class, and join Channel
Groups after the parent connect process has completed.