Use YAML instead of JSON for tslint config
In setting up my new editor, I got a bunch of error highlights for the comments
in tslint.json; rightly so, because the JSON spec does not allow comments. But,
comments are useful. So I moved the file from JSON to a more comment-friendly
YAML format. Also adding some additional configs silencing lint errors for
multiple classes in a single file and use of non-arrow functions.