Correct handling of missing partId from ICE
ICE only includes partial information on objects in the parents
field of an Entry, and excludes specifically type-specific fields
e.g. strainData, and excludes partId. The conversion code was
warning of a missing type-specific field by doing a direct lookup
on the translated part_id field, resulting in a KeyError.