Split database utilities into own module
The existing utilities module mixed together functions using the database with
lower-level functions that do not access the database. Doing so introduced
complications when model-layer code would import the utility functions, where a
Model class could depend on a utility function that in turn depended on the
Model class being loaded. Splitting utilities into a query module that does
database access, and a utilities module that does not, makes it clearer where
the abstraction boundaries lie.