Restructure Jenkinsfile to handle email errors
The pipeline definition was previously sending an email in a finally block that
also was cleaning up Docker resources. If the email failed (generating an
exception), none of the cleanup would happen. Move the email code to the very
end of the file, so failures there do not affect the rest of the stages.