Fix auto-validation of emails from LDAP
The LDAP backend code has a hook so that any email address coming from LDAP
gets auto-verified, so that users do not need to receive an email and click on
a verification link. This code was using update_or_create() to add an entry in
the database for the email, but because it was using a filter of email__iexact,
it would end up always trying to insert an empty email address. The
documentation on update_or_create() points to get_or_create() to describe how
filters are handled, and the algorithm described excludes any filters that are
not direct fields. The change here modifies the validation hook to properly set
the email from LDAP, and cleans up edge-cases.