Remove vestigal export option
See EDD-451. The perl version used the option to specify whether each row would
be a single point, or if multiple points for the same measurement would get row
with additional columns per X value. It would also allow selecting a count of
the number of points with the time range of measurements. The option for
row-per-point or row-per-measurement moved to the layout selection, and the
"summary" option with time range and count of values is not useful.