Prevent websocket connections on anonymous session
Refactoring the code around the notification websocket so that no connections
are attempted as long as no user is currently logged in. This is determined by
looking for the "inbox" menu in the base template, and only building the
websocket connection when it is present. The existing exported
NotificationSocket is created in "stub" mode, where clients can use all the
provided APIs, but as no underlying connection exists, no messages are
sent or received.