Explicitly load TinyMCE theme styles
The default behavior of TinyMCE v5.x is to have the tinymce.init() function
dynamically load in styles for the editor. These files will not be in place
when using webpack, so we must explicitly load the relevant styles in Webpack,
and pass in options to tinymce.init() to disable the dynamic loading behavior
in favor of the bundled-and-minified Webpack stylesheet.