Stage 4 of module reorg ... --> edd/load
Moving bits of code relating to loading in measurement values to new Django app
`edd.load`. This commit is limiting the changes to only the older "table" form
of measurement loading. A later commit should handle moving the newer "wizard"
form of loading from `edd_file_importer` to the same app. The
`edd_file_importer` app already had substantial dependencies on code in the new
app module, which is why half the changed files are in `edd_file_importer`;
these changes are only to re-point dependencies to the new location.