Allow Jenkinsfile to use core dockerignore
Because the Jenkinsfile build was piping the Dockerfile to STDIN, there was no
way for Jenkins to take advantage of dockerignore. So, instead of piping to
STDIN, change the Dockerfile to not require edits. Then the build can make use
of the new dockerignore. To do this, I moved both the node and core build
scripts out of strings in the Jenkinsfile, and into new shell scripts in
./bin/jenkins, then have the Jenkinsfile run those scripts.