Fix import statements to reference true dependency
Several parts of the edd_file_importer tests were importing dependencies by
depending on the dependencies of unrelated modules. Replacing these transitive
dependencies with imports to the actual dependency source, instead.

As an example, if you have modules A.py and B.py, both depending on code in
module C.py as below, then the correct way for both A and B to import is from
C, not from each other.

    # C.py
    def shared_dependency():
        pass

    # A.py
    # this is the way
    from .C import shared_dependency

    # B.py
    # should not do this!
    from .A import shared_dependency

It is a bad practice to use this transitive dependency importing, because by
doing so, the implementation of e.g. module B becomes more tightly coupled to
the implementation of module A. Should module A update in the future to no
longer depend on C.shared_dependency, then module B will break!