Add approved flag to profiles
Extends the Profile app to include an `approved` flag on profiles. When paired
with the manual verification backends
(`edd.auth_backends.ManualVerificationModelBackend` or
`edd.auth_backends.ManualVerificationAllauthBackend`), any accounts that are
not approved will fail login attempts, with a message loaded from setting
`EDD_APPROVAL_MESSAGE`, or a default message stating to contact
`EDD_APPROVAL_CONTACT`, or failing that, a default message stating to contact
the `SERVER_EMAIL` address.