Copy exception structure from edd_file_importer
All the Exception types defined in `edd_file_importer` copied into parallel
structure in `edd.load`, and merged with the existing exceptions. This is a
first step to getting the `edd_file_importer` functionality moved into
`edd.load`. All the error reporting code mixed in will be copied to a
separate module.

Copying at least some of the Exception types is a prerequisite for copying the
models, so going ahead and copying them all.