Copy parsers from edd_file_importer
Note that while the tests are mostly the same, they have been restructured. The
big changes are getting rid of the comparisons of ParseResult objects with
ParseResult objects created from JSON run through a custom deserializer. It's
not useful to test one parser vs the result of another parser, especially as
the second (JSON custom deserializer) parser only appears to be used in tests.
Instead, the results are tested vs a long list of assert statements. Wherever
possible, we want to eliminate logic/branching in tests, because that's where
bugs live, and bugs in tests makes everyone's life miserable.