Create LoadRequest object to track data loading
The LoadRequest is an analog to the Import model class in `edd_file_importer`.
The data structure is used in a transient / semi-permanent way in the
`edd_file_importer` code, yet is stored permanently in a RDBMS. This analog is
stored in Redis with an expiration built-in. It's also dropping some fields
from the model class that were never used (description, created, updated,
name), and dropping fields that were included due to incorrect assumptions
(category, file_format).

This covers the second half of EDD-1272, to overhaul the import models.