Add resolver package to edd.load
The code in the new resolver.py file is taken liberally from code in
edd_file_importer/importer/table.py, limited to just the classes responsible
for translating tokens from parsers into database IDs. The high-level structure
of the copied classes is the same, but is heavily altered internally, as the
internals made multiple references to an Import model that does not exist.

Some further refactoring should happen, especially around migrating lookups of
line/assay names and unit names, respectively, to their own classes, as with
TypeResolver for MeasurementType lookups. Also, ImportCacheCreator is getting
created with most of the properties of the ImportResolver that creates it, and
should instead just get a reference to the ImportResolver and access the data
it needs from a defined API.