Remove inline script from study HTML template
This has been in there a very long time. An inline script that existed only to
set a value for the current study primary key, which gets used in a few places
in the current TypeScript frontend. Now the script is gone, and the only places
that use it now get the value from the AJAX call that populates the EDDData
global object. I couldn't find a ticket for this one, but it's been a TODO for
*years* now, and I saw I had this fix stashed.