Upgrade JQuery type definitions to 3.x
We've been using JQuery version 3.x for some time, but were still on the 2.x
type definitions. The biggest difference is the 3.x definitions are much more
explicit about the return value of jQuery().val(); where before the definitions
essentially returned `any`, now it specifies `string | number | string[]`, as
technically a `<select multiple="multiple">` input type can return an array of
strings. The docs also mention the possibility of returning a number, but this
does not appear to happen (even with `<input type="number">`). So now the code
must do `$(selector).val() as string` to make it explicit that we're not
dealing with a multiple select.

Pinning this to 3.3.35, because 3.3.36 introduces a regression where some
JQuery methods end up getting type mismatches between JQuery<Element> and
JQuery<HTMLElement>.

https://github.com/DefinitelyTyped/DefinitelyTyped/issues/44268