Fix comparison error in multi-line edits
When determining how to edit multiple lines, a comparison is done between all
the lines to see if a particular field has the same value across all edited
lines. When comparing array-like fields, there could be errors where an
array-like value gets compared to an undefined value. Using a Set to do the
comparison instead, and removing the possibility of having undefined values.