Fix tests failing after EDD-1285 changes
A few tests started failing after previous change for EDD-1285 in commit
c7b8016cca456b4aeaa8ef5409b3f886190bfcce. The issue was in changing the
response on export download requests to a HTTP 204 NO CONTENT, which does not
happen in the current UI, but could potentially happen and is covered in tests.
The 204 response should only happen for explicit download requests, while other
requests should reload the page with the updated form values.

There were also problems in validating the no-input fallback cases for export
selections. In some cases, the selection keys would exist with all empty
values. So it was not sufficient to check that they keys did not exist, the
values also needed checked.