Cleanup interactions between frontend and backend
The template HTML for legacy import was using code similar to the export code,
which supports multiple study exports. Import only goes to a single study, and
the multiple study values did not exist, so the title would always appear as
"Importing to 0 studies".

Fixed handling of messages in final wizard step. Add in hooks to acknowledge
and dismiss warnings occuring during use of data loading wizard. Once all
warnings are dismissed, the wizard step can validate and move on to next step.

Using some better conditions on automatic actions and transition requests in
the wizard. Without the conditions, going back and forth between steps could
result in unexpected behaviors.