Fix color selection when filtering out-of-order
When filtering lines in the graph view out-of-order, the color assignment was
happening in the order of items in the filter list, but any items already
having an assigned color would keep that color. Because of this, any lines that
were already selected could get their colors prematurely assigned to earlier
items in the list. Instead, doing one pass to find items missing a color
assignment, then a second pass to do the actual assignment of the next
available color.