Fix locking timeout during startup
Some parts of container startup are behind locks, because the container is
accessing and modifying resources that are shared between containers. These
locks were not added with timeouts, so when container orchestrator would kill a
container in the middle of the lock, the system would become deadlocked. Adding
in the timeout should resolve the problem.