Remove the assay filter from Data view
The Data view in EDD was confusing because of having both a filter for lines
and for assays. Here the assay filter is removed, leaving only a filter by line
name, plus the other metadata for lines and assays.

Also removing sections of dead code, and a lint error in the graphing module.