Fix user lookup when login with email
The authentication backend for LDAP would resolve either username or email to a
DN for authn, but would then always use the "login" field to resolve as a
Django username. When a user tries to authenticate with the other value in the
"login" field, LDAP continued to resolve, but Django would end up creating a
new User record, because username != email. The change here short-circuits
this, by taking advantage of the LDAP backend always creating an EmailAddress
record for LDAP logins. Now, if someone first logs in with an email, then logs
in with a username, the backend will check the EmailAddress record from the
LdapUser record first, before trying to create a duplicate account.