Add coverage report to dev Docker images
A custom script is added to only dev builds of the EDD Docker image, which runs
tests with pytest under coverage, then prints a report. This was previously
done in Jenkins scripts, but now it's built-in to the image. In addition, the
new test running script will write a JSON-formatted report.

The JSON report makes it easier to integrate coverage information in text
editors capable of reading and displaying it. Outputting a JSON report also is
the first step in generating a "Code Insight" report to display the same
information in Bitbucket. With a few more additions, the Jenkins test runner
script can modify the output, then call back to Bitbucket to display coverage
information inline in pull requests, etc.