Refactor API to access "system" metadata
Deprecating the previous method of looking up UUIDs in the SYSTEM_META_TYPES
attribute. Instead, call a system() class method on MetadataType with a system
type name. The classmethod will use get_or_create(), so there is always a
record for the type.

The system types are also limited to those types that are referenced in code
for specific functionality and types that refer to model fields and are exposed
via metadata APIs. Had to do some minor refactoring in edd.describe tests that
assumed some types not included in those criteria already existed.