Squash main app migrations
The aggregated changes to the schema for the main app has a good amount of
cruft and churn when running through the existing migrations. The change here
is created by moving aside the existing migrations and re-running `./manage.py
makemigrations main`; then copying over the bootstrap data changes.

Having the squashed migration run, instead of the older 22 files, saves about
25 seconds -- or about 25% -- off setup time for running tests on my machine.