Add support for applying squashed migrations
The startup invoke script detects pending migrations by running the
`showmigrations` Django command. The output of this will show squashed
migrations as already applied if all the migrations it replaces are applied;
however, the database log will not have the squashed migration applied. To
account for this, always run `migrate` command with the `--fake-initial` flag
when there are no pending migrations shown. This will ensure that the squashed
migration gets added to the migration log, and allows for the eventual removal
of the deprecated migrations.