Fix notification tests
Upgrading the libraries used to communicate channels to Redis exposed some
problematic test setups. Tests were creating users with a random username, but
all had a `None` primary key. This in turn meant that all the tests were using
the same "channel" for multiple users, as channels are named in part with user
primary keys. Moving the creation of test users to a pytest fixture, with a
randomized primary key generated (but not saved to database), resolves the
test failures.