Replace Lines table library with Handsontable
From our custom DataGrid library. The configuration for DataGrid was too
convoluted and onerous to continue using, and was blocking progress on handling
display mode for viewing Line replicates. The updated script for that page is
greatly simplified, and switching to a replicate view mode can be done in a
handful of extra lines.