Remove customized middlewares
Neither of the customized middlewares are required anymore. The subclass of the
ThreadLocalMiddleware was a patch to get around a deprecated Django API, which
has since been fixed upstream. The EDDSettingsMiddleware was adding an
attribute to requests for display in templates, but has since been superceded
by the edd.branding app adding a custom templatetag instead.