Clean up query module with front-end
Long overdue. The front-end displays data from EDD starting with a query to
fetch an EDDData object. This object was originally overkilling on the data
provided; e.g. listing the entire roster of users in the system. Changes here
serve to both limit the provided data to relevant records, and improve the
queries providing that data. The total number of queries is reduced, the
execution time is roughly half, and prefetch queries that filter on potentially
large numbers of IDs are eliminated.