Add support for using S3 storage backend
This is a necessary step for moving toward multiple-node setups of EDD, where
we cannot guarantee that nodes will have access to the same filesystem. Using
NFS for sharing a filesystem does not play well with Django, as it makes use of
filesystem locking. Locking through the filesystem is fraught enough, and doing
so over NFS only compounds the problem. Using a storage backend with the S3 API
solves these issues, as the locking mechanics are handled in the external
storage service.

As a practical matter, this involves removing the assumption of working with a
local filesystem to the EDD startup script, adding a configurable storage
backend for the `edd.load` app over simply re-using the default upload backend,
and offering an admin task to migrate from previous storage to the S3 storage.