Add swarm-aware scripting to Jenkins deploy
Because the Jenkinsfile can now deploy EDD to a Docker Swarm, the script to
launch EDD needs some modifications to ensure that the launch, testing, and
teardown work properly. Running `docker exec` will only work on the specific
node running a container, so must constrain those containers that the script
will exec to the manager node launching the stack. Checking on health of all
other containers must work through to the service and task, as healthcheck info
is also only available on the specific node running a container.