Add support for add/remove replicate per line
View additions and the tests for those additions, allowing for the UI to
manually select lines to group as replicates, or to remove a replicate grouping
from lines. This allows for converting older studies, or studies that did not
make use of the replicate feature during the line definition phase, to using
replicates. This also makes it easier to quickly validate behavior in the UI
when switching between individual display and replicate display modes.