Combine buttons on Description view
Preparing to add more control elements (buttons) to manage setting and
un-setting replicates. The existing view had eight buttons along the bottom of
the table, and the buttons only had enough horizontal space on wide screens --
a 768 pixel wide window would cause the buttons to overflow space and stagger
in a strange way. Changing replicate status would make it ten buttons, and
stretch the screen real estate required even further.

The new layout has three basic button functions, with additional functions in a
dropup menu: Edit, Add, and Export. The horizontal space required for only
three discrete block elements in the page is much smaller than eight or more,
and now comfortably fits on smaller screens or windows.

Also cleaning up the display of other elements on the page. Using a jumbotron
element for the loading and no-lines messages. Replacing cloned controls for
overlay while scrolling, with a constrained table view that does not need a
separate set of controls with a fixed-position style. Removing export modal in
favor of the dropup menu.