Fix rendering of strains w/o URLs
Strains now are enforced to have a registry_url property with a link to e.g.
ICE for that strain. There are some older strain records that did not have this
requirement, and that will render as broken links with the previous
implementation of the renderer. Now those link-less strains will render as just
the strain name, without a hyperlink.