Migrate from npm to yarn for package installs
The yarn node package manager has a built-in feature to resolve transitive
dependencies to specific versions. The dependencies for source-map-resolve and
chokidar from webpack's sub-dependencies were emitting deprecation warnings.
Forcing these transitive dependencies to more recent versions reduces the total
number of installed dependencies (from 617 to 486), and eliminates warnings.

Eliminate dependency listings that are not directly used: "fibers",
"node-sass", "qtip2". The dependency on "promise" is also not used, but is an
undeclared dependency for "react-stepzilla", so is kept here.