Refactor data view AJAX queries
The frontend for the data view is now no longer constructing AJAX URLs in the
TypeScript code, and is instead getting URLs directly from templates and/or
other AJAX calls. This brings the ticket for eliminating frontend URL
construction (EDD-1261) a bit closer to completion.

Simplified the AJAX view for fetching graph data, so that instead of two nearly
copy-pasted view functions, there is now only one. The return payload is now
typed as well on the frontend TypeScript side.