Change templates so new import is the default
Similar change to what was done with master / 2.7.x branch, but without the
wholesale refactoring and replacement of the import code. The template now
links to the pre-refactor version of import as the default mode, and the
link to the import UI from EDD's Perl era is now a button in the updated
import. Click once, get new import. Click again after page load, get the
legacy import.