Fix healthcheck for Celery 5.0+
Similar to fix to startup invoke script, the default healthcheck for the Celery
worker container(s) was using 'global' flags after the sub-command name,
instead of before the sub-command. Fixes the healthcheck definition so that the
`-A` and `--workdir` global option flags and arguments come before the
`inspect ping` sub-commands.