Update available buttons in Data view
Simplifying the available action buttons on the Data view. Where there were six
buttons on the "bottom" (Edit Assays, Edit Measurements, Add Measurement,
Delete, Restore, Export Data), now there are three (Edit, Add, Export Data).
For the moment, this is an intentional step back in functionality, dropping
ability to edit measurements, and delete or restore records. Deletion and
restoring records should go in a drop-up menu on the Edit button, similar to
the Lines/Description view. Editing measurements should function to the change
in the Export Data button, collecting the selected records from the table and
submitting a <FORM> directly. This should happen along with a refactor to make
the Django View for editing measurements its own view, instead of embedding it
in the StudyDetailView.