Remove "export" subtask from StudyLinesView
Similar to change in previous commit, the export functionality in the UI for
the Lines / Description view is now directly calling to the dedicated export
view(s), and not submitting a POST to the same page with an instruction to
delegate to the export view. Moved the tests that were covering the delegation
to run in the edd.export module.