Add a static files refresh script
This helper script assists in development on the front-end code. Running the
collectstatic command will copy files to the right place, but Django will still
believe that its staticfiles manifest will be the previous version. Sending a
HANGUP signal to the main process will force Django to use the updated
manifest, with the updated files.