Switch to eslint-webpack-plugin from eslint-loader
Resolves EDD-1300.

The eslint-loader package is deprecated as a way of running eslint on files
during build. It is replaced with a plugin instead, eslint-webpack-plugin, with
simpler configuration. To get everything to work, typescript and eslint also
need updating.

The newer eslint comes with some additional lint rules that resolve as errors,
and are fixed in this commit. There are several warnings as well, nearly all
around making explicit the types used as the public interface to a module.
Fixing these warnings would result in a substantial amount of code change, so
defering the fixes until later.