Remove multiprocessing from init-config invoke
Following the addition of some debugging code to the EDD config container
invoke script in commit fe6f8db5be97a806ef9004dd953d10f0d066253b, I finally
caught some clues to why Jenkins will sometimes fail builds in the init step.
Now that it's printing some progress information, Jenkins can show more closely
where the process is failing. The output on the failed run was:

    Configuring postgres
    Configuring rabbitmq
    Configuring redis
    Configuring mail
    Configuring solr
    Configuring nginx
    Configuring ice
    Configuring core
      yq w -s /tmp/http.script -i service/http.yml
      yq w -s /tmp/websocket.script -i service/websocket.yml
      yq w -s /tmp/worker.script -i service/worker.yml
      yq w -s /tmp/postgres.script -i service/postgres.yml
      yq w -s /tmp/rabbitmq.script -i service/rabbitmq.yml
      yq w -s /tmp/ice.script -i service/ice.yml

The lines reading "Configuring ..." are prints, and the "yq" lines are shell
commands executed by the config invoke script. The script was building a list
of individual services to configure, then was running them in subprocesses
using multiprocessing.Pool. It looks like the problem is the Pool object is
using map, but will sometimes fail to block until all the items passed into
map() finish processing. There should be lines like the following once all the
yq scripts are run, but they are missing:

      yq m -a service/http.yml service/websocket.yml service/worker.yml \
          service/postgres.yml service/rabbitmq.yml service/ice.yml \
          service/smtp.yml service/redis.yml service/solr.yml \
          > /tmp/edd-config/docker-compose.override.yml
      tar -czf /tmp/edd-config.tgz -C /tmp/edd-config .

The multiprocessing module was included originally because of large amounts of
latency when running the script, in an attempt to divide-and-conquer the work.
The latency turned out to be because passing data to a shell subprocess
standard input directly, would write to the shell command one byte at a time,
with a sleep between each byte. Rather than trying to access standard input,
all subshells are instead writing out data to temporary files, and passing the
temporary file in the subshell command. With the latency gone, there's no
longer a need for multiprocessing, and it's apparently causing a problem. So
now, it's gone.