Remove deprecated EDD_USE_PROTOTYPE_IMPORT refs
The prototype is re-written, and is now the default. The use of this setting is
no longer supported, and the code guarded by the setting no longer exists.

Updating the URLConf around Channels / websockets near where the setting was
used, to use the updated API from the `django.urls` module, instead of the old
API from `django.conf.urls`.