Refactor FileDropZoneHelpers to new module
The DescriptionDropzone module is replacing FileDropZoneHelpers, with a more
appropriately namespaced collection of functions. The module is a set of
callback functions to use with the wrapped FileDropZone class in the Utl
module. Previous code was lacking in type information, and the subroutines used
by the class/module had several awkward boundary conditions hidden in its
control flow. For example, the replaced code would only display the buttons to
handle problems communicating with the strain registry if that error is the
first returned; then, all subsequent errors are also treated as errors
communicating with the strain registry.

The updated module adds in typing information, and docstrings describing the
functions and data structures where it seemed appropriate to include them.