Remove color from un-selected lines in Filter
Small refactoring to better handle updating the Filter display between changes
in selected items. Each FilterValue record originally had a "hidden" property,
which is now renamed to a more appropriate name for its intended function, of
"dimmed". Now after the Filter selects Items for display, any Lines that are
not currently shown have their color removed.

A future update could use this value to apply further styling to the inputs in
the Filter, such that values that have currently displayed Items are visually
distinct from values that are removed by the Filter.