Refactor edd.notify async tests
Part of troubleshooting and fixing EDD-1318, where WebSocket connections to
edd-test.jbei.org are not properly authenticating, and are immediately closed.

The tests are replacing the previous, hacky way to "login" by specially
crafting a scope object sent to the Consumer class. Instead, use the Django
Test Client to do a force_login() call with the test user, and then initialize
the WebsocketCommunicator object with the HTTP headers that a real client would
send along -- particularly the cookies. This is similar to the approach
described in this GitHub issue comment in the channels project:
https://github.com/django/channels/issues/903#issuecomment-365448451

Also taking the opportunity to refactor a handful of scattered calls to the
sync_to_async decorator. Each call now has its own async alias function, so the
Consumer code reads more like typical async code.