Fix template for expired email confirmation tokens
The confirm email template was always rendering a FORM element to confirm an
email, even if no current valid token existed. Because there is no valid token,
the URL in the action attribute of the FORM element does not exist, and
generates a 500 error.

Instead, skip rendering the FORM element when no valid confirmation object can
be found, similar to structure of upstream template.