Fix flickering select-all checkbox
The Handsontable library used to display tables of EDD data will frequently
redraw table headers via replacing the DOM nodes for those headers. There was a
checkbox explicitly rendered as a select-all control for the line, assay, and
measurement table views, to handle selecting or de-selecting all items shown in
the table. This worked, but the UI would show a flicker whenever the table
state changed, due to the Handsontable library frequently replacing the input
and its current state, with a new checkbox having cleared state.

This code is now refactored, so the control to toggle selection states is added
to the DOM outside of the table during table setup. Since the control is no
longer part of the headers, there's no problems with the control getting
replaced and showing a flicker in the UI. With the control existing outside of
the table, some additional code is required to position the control as if it
were in the table headers.

As an additional side effect, more of the code dealing with configuring options
for Handsontable is moved out of the page scripts, and is instead kept in the
table/Config.ts module. This improves the code by concentrating almost all the
options for Handsontable to the Config module, leaving only a few to be set in
the page scripts.