Fix initialization of Dropzone elements
This is a strange error. Everything works fine if run locally. In the deployed
version, for some reason the TypeScript code for `new Dropzone()` throws a
TypeError of "x is not a constructor", where "x" varies depending on the
specific script loaded. Importing the dropzone module also adds Dropzone to the
window scope; so, this fix references the window scope of the class, instead of
whatever TypeScript compiles.

Note, that the content of the scripts deployed and the version running locally
are identical. It's unclear why there is any difference in behavior.