Pin nodejs dropzone package to 5.7.2
Recent issues with initializing Dropzone elements, where browsers throw a
TypeError stating that the (minified name of) Dropzone is not a constructor,
appear to be related to several version bumps from the dropzone project. The
previous code was using 5.7.2; the current version is 5.7.6, and the author
makes note of some fairly major refactoring. Trying to revert to the previously
working version, until the issue of getting the project imported to our
environment is resolved.