Update documentation with version migration info
Adding additional information to the documentation on migrating an EDD
deployment to the new 2.7.0 version. Expanding on the existing documentation
around using an S3 storage backend, with the specific commands to run to set up
the storage bucket(s). Adding another section on properly setting up the
configuration for imports.