Remove call to non-existant method
The refactoring for the notification async tests, had a call erroneously added
to the connect method for import notifications. The call exists on the general
notification consumer, and is meant to populate existing notifications on
connect. The import notifications are a per-page notification, and do not
support having pre-existing notifications to display on connect.