Fix attach of click handler to permission button
This was a long-standing annoying bug on Firefox, but worked fine on Chrome and
Safari until very recently. It appears that setting the click handler on the
permission button was failing, because the function that added the handler was
wrapped in a page load event handler, which in turn was wrapped in a handler
that waited for the EDD AJAX data for the study to download. The page would
finish loading before the AJAX custom event fired, and thus the handler that
attached the click event would not run.

From a quick scan, it appears there is no longer any code in the setup for
permissions that requires waiting for the AJAX payload to finish downloading
and processing. Moving the permission setup to the top level of the script,
next to the custom event, fixes the problem. It seems that Firefox was first to
stop processing handlers for a load event if the event handler is registered
*after* the page already loaded. Chrome and Safari were continuing to
automatically trigger the handler if the page already loaded until recently.