Switch custom profiles UserManager base class
Instead of using a plain Manager object to attach the profiles custom manager
to the User model, extend the default UserManager instead.

I noticed that some code paths would end up calling the profiles manager as if
it were a UserManager, and would throw AttributeError exceptions, because the
plain Manager is not a UserManager. Changing the base class should resolve
that problem.