Override Bootstrap 3 style for DETAILS / SUMMARY
The normalize stylesheet included with Bootstrap 3 includes adding a
`display: block` property to SUMMARY HTML elements. However, doing so removes
the triangular "twisty" icon indicating the open / closed state of the parent
DETAILS element of the SUMMARY element.

Adding a rule for any SUMMARY element that is a child of an HTML element (which
should be all of them), forcing the `display: list-item` property, and has a
higher priority than the rule set by Bootstrap. With this addition, the icons
once again show up, indicating the element can be expanded / collapsed.